Add custom security group
The security group in the nested stack changed to only allow ssh inbound form the SshSecurityGroup. The jumphosts need to allow ssh from anywhere and therefore need a custim security group.

NB: This also has the effect of removing the port 80 access to the jumphosts, which should not be an issue.